requirements.txt
hither
scipy
python-slugify
spikeextractors
numpy
pynwb
